--- src/main/java/com/auction/E_Commerce_Auction/ECommerceAuctionApplication.java
@@ -1,13 +1,0 @@
-package com.auction.E_Commerce_Auction;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ECommerceAuctionApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ECommerceAuctionApplication.class, args);
-	}
-
-}
