+++ src/main/java/com/auction/ecommerce/ECommerceAuctionApplication.java
@@ -1,13 +1,13 @@
-package com.example.auction_ecommerce;
+package com.auction.ecommerce;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AuctionEcommerceApplication {
+public class ECommerceAuctionApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(AuctionEcommerceApplication.class, args);
+		SpringApplication.run(ECommerceAuctionApplication.class, args);
 	}
 
 }
