--- src/test/java/com/example/auction_ecommerce/AuctionEcommerceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.auction_ecommerce;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuctionEcommerceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
